+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next Up
 
+
+
 # 6.6.23
 
 - Improve speed of class muncher by pre-filtering sources for subclasses.